Update TopBar.qml
To raise mirage's window in a stacking wm clicking on an area without text, especially in the topbar, is probably what most users do. In the case of the 2nd button in the topbar this currently accidentally opens a browser window and thus becomes a privacy issue. So for now require a double click to open mirage's homepage. Alternatively the size of this button should at least be closely restricted to the text "Mirage ${version number}". In the long term it might be preferable to put this link into some kind of "about" section instead of the "title" section.