AddChat tab titles: say "group" instead of "room"
Helps users new to Matrix in knowing what to click, as the direct chat
option explicitely creates a 1-to-1 experience, while "create room"
(now "create group") could be also be used as such but asks for more
details and wouldn't mark the room as direct.

The content of the tabs still say "room" as that is the correct general
term for anything that contains one or more user.